api/requirements.txt
fastapi>=0.110,<0.115
uvicorn[standard]>=0.27,<0.30
SQLAlchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0
PyJWT[crypto]>=2.8,<3.0
requests>=2.31,<3.0

